Reorganize the project
Organize the classes.
